+++ app/upgrades_test.go
@@ -8,6 +8,7 @@ import (
 	"sort"
 	"strings"
 	"testing"
+	"time"
 
 	"github.com/rs/zerolog"
 	"github.com/stretchr/testify/suite"
@@ -15,9 +16,13 @@ import (
 	"github.com/tendermint/tendermint/libs/log"
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 
+	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
 	"github.com/cosmos/cosmos-sdk/server"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
+	"github.com/cosmos/cosmos-sdk/x/bank/testutil"
+	"github.com/cosmos/cosmos-sdk/x/staking/types"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 
 	msgfeetypes "github.com/provenance-io/provenance/x/msgfees/types"
 )
@@ -29,6 +34,8 @@ type UpgradeTestSuite struct {
 	ctx sdk.Context
 
 	logBuffer bytes.Buffer
+
+	startTime time.Time
 }
 
 func TestUpgradeTestSuite(t *testing.T) {
@@ -53,7 +60,25 @@ func (s *UpgradeTestSuite) SetupSuite() {
 	defer SetLoggerMaker(SetLoggerMaker(bufferedLoggerMaker))
 	s.app = Setup(s.T())
 	s.logBuffer.Reset()
-	s.ctx = s.app.BaseApp.NewContext(false, tmproto.Header{})
+	s.startTime = time.Now()
+	s.ctx = s.app.BaseApp.NewContext(false, tmproto.Header{Time: s.startTime})
+}
+
+// GetLogOutput gets the log buffer contents. This (probably) also clears the log buffer.
+func (s *UpgradeTestSuite) GetLogOutput(msg string, args ...interface{}) string {
+	logOutput := s.logBuffer.String()
+	s.T().Logf(msg+" log output:\n%s", append(args, logOutput)...)
+	return logOutput
+}
+
+// LogIfError logs an error if it's not nil.
+// The error is automatically added to the format and args.
+// Use this if there's a possible error that we probably don't care about (but might).
+func (s *UpgradeTestSuite) LogIfError(err error, format string, args ...interface{}) {
+	if err != nil {
+		args = append(args, err)
+		s.T().Logf(format+": %v", args)
+	}
 }
 
 // AssertUpgradeHandlerLogs runs the Handler of the provided key and asserts that
@@ -63,20 +88,16 @@ func (s *UpgradeTestSuite) SetupSuite() {
 // entries in expInLog are expected to be full lines.
 // entries in expNotInLog can be parts of lines.
 //
-// This returns the log output and whether all assertions passed (true = all passed,
-// false = one or more problems were found).
-func (s *UpgradeTestSuite) AssertUpgradeHandlerLogs(key string, expInLog, expNotInLog []string) (logOutput string, allPassed bool) {
+// This returns the log output and whether everything is as expected (true = all good).
+func (s *UpgradeTestSuite) AssertUpgradeHandlerLogs(key string, expInLog, expNotInLog []string) (string, bool) {
 	s.T().Helper()
-	defer func() {
-		allPassed = !s.T().Failed()
-	}()
 
 	if !s.Assert().Contains(upgrades, key, "defined upgrades map") {
-		return // If the upgrades map doesn't have that key, there's nothing more to do in here.
+		return "", false // If the upgrades map doesn't have that key, there's nothing more to do in here.
 	}
 	handler := upgrades[key].Handler
 	if !s.Assert().NotNil(handler, "upgrades[%q].Handler", key) {
-		return // If the entry doesn't have a .Handler, there's nothing more to do in here.
+		return "", false // If the entry doesn't have a .Handler, there's nothing more to do in here.
 	}
 
 	// This app was just created brand new, so it will create all the modules with their most
@@ -88,40 +109,174 @@ func (s *UpgradeTestSuite) AssertUpgradeHandlerLogs(key string, expInLog, expNot
 	// any old state for that module since anything added already was done using new stuff.
 	origVersionMap := s.app.UpgradeKeeper.GetModuleVersionMap(s.ctx)
 
+	msgFormat := fmt.Sprintf("upgrades[%q].Handler(...)", key)
+
 	var versionMap module.VersionMap
 	var err error
 	s.logBuffer.Reset()
 	testFunc := func() {
 		versionMap, err = handler(s.ctx, s.app, origVersionMap)
 	}
-	if !s.Assert().NotPanics(testFunc, "upgrades[%q].Handler(...)", key) {
-		return // If the handler panics, there's nothing more to check in here.
+	didNotPanic := s.Assert().NotPanics(testFunc, msgFormat)
+	logOutput := s.GetLogOutput(msgFormat)
+	if !didNotPanic {
+		// If the handler panicked, there's nothing more to check in here.
+		return logOutput, false
 	}
-
-	// Add the handler log output to the test log for easier troubleshooting.
-	logOutput = s.logBuffer.String()
-	s.T().Logf("upgrades[%q].Handler(...) log output:\n%s", key, logOutput)
 
 	// Checking for error cases should be done in individual function unit tests.
 	// For this, always check for no error and a non-empty version map.
-	s.Assert().NoError(err, "upgrades[%q].Handler(...) error", key)
-	s.Assert().NotEmpty(versionMap, "upgrades[%q].Handler(...) version map", key)
+	rv := s.Assert().NoErrorf(err, msgFormat+" error")
+	rv = s.Assert().NotEmptyf(versionMap, msgFormat+" version map") && rv
+	rv = s.AssertLogContents(logOutput, expInLog, expNotInLog, true, msgFormat) && rv
 
-	logLines := strings.Split(logOutput, "\n")
+	return logOutput, rv
+}
+
+// ExecuteAndAssertLogs executes the provided runner and makes sure the logs have the expected lines and don't have any unexpected substrings.
+func (s *UpgradeTestSuite) ExecuteAndAssertLogs(runner func(), expInLog []string, expNotInLog []string, enforceOrder bool, msgFormat string, args ...interface{}) (string, bool) {
+	s.logBuffer.Reset()
+	didNotPanic := s.Assert().NotPanicsf(runner, msgFormat, args...)
+	logOutput := s.GetLogOutput(msgFormat, args...)
+	if !didNotPanic {
+		return logOutput, false
+	}
+
+	rv := s.AssertLogContents(logOutput, expInLog, expNotInLog, enforceOrder, msgFormat, args...)
+
+	return logOutput, rv
+}
+
+// splitLogOutput splits the provided log output into individual lines.
+func splitLogOutput(logOutput string) []string {
+	rv := strings.Split(logOutput, "\n")
+	if len(rv[len(rv)-1]) == 0 {
+		rv = rv[:len(rv)-1]
+	}
+	return rv
+}
+
+// AssertLogContents asserts that the provided log output string contains all expInLog lines in the provided order,
+// and doesn't contain any string in the expNotInLog slice.
+// Returns true if everything is as expected.
+func (s *UpgradeTestSuite) AssertLogContents(logOutput string, expInLog, expNotInLog []string, enforceOrder bool, msg string, args ...interface{}) bool {
+	s.T().Helper()
+	rv := s.assertLogLinesInOrder(logOutput, expInLog, enforceOrder, msg+" log output", args...)
+	rv = s.assertLogDoesNotContain(logOutput, expNotInLog, msg+" log output", args...) && rv
+	return rv
+}
+
+// assertLogLinesInOrder asserts that the log output has whole lines matching each of the expInLog entries,
+// and that they're in the same order (allowing for extra lines to be in the log output that aren't in expInLog).
+// Designed for AssertLogContents, please use that.
+func (s *UpgradeTestSuite) assertLogLinesInOrder(logOutput string, expInLog []string, enforceOrder bool, msg string, args ...interface{}) bool {
+	if len(expInLog) == 0 {
+		return true
+	}
+
+	logLines := splitLogOutput(logOutput)
+
+	allThere := true
+	// First, just make sure all the lines are there.
+	// This gives a nicer failure message than if we try to do it while checking ordering.
 	for _, exp := range expInLog {
 		// I'm including the expected in the msgAndArgs here even though it's also included in the
 		// failure message because the failure message puts everything on one line. So the expected
 		// string is almost at the end of a very long line. Having it in the "message" portion of
 		// the failure makes it easier to identify the problematic entry.
-		s.Assert().Contains(logLines, exp, "upgrades[%q].Handler(...) log output.\nExpecting: %q", key, exp)
+		if !s.Assert().Containsf(logLines, exp, msg+"\nExpecting: %q", append(args, exp)...) {
+			allThere = false
+		}
+	}
+	if !allThere {
+		return false
+	}
+	if !enforceOrder {
+		return true
 	}
 
+	// Now make sure they're in the same order, allowing for extra lines in the log that might not be expected.
+	e := 0
+	for _, logLine := range logLines {
+		if expInLog[e] == logLine {
+			e++
+			if e == len(expInLog) {
+				return true
+			}
+		}
+	}
+
+	// We didn't get to the end of the expected list. Issue a custom failure.
+	failureLines := []string{
+		"Log lines not in expected order.",
+		fmt.Sprintf("End of log reached looking for [%d]: %q", e, expInLog[e]),
+		// Note: We know that all expInLog lines are in the log output.
+		// So the above loop will find at least the first entry, and e will be at least 1.
+		fmt.Sprintf("Expected it to be after [%d]: %q", e-1, expInLog[e-1]),
+	}
+	// Putting the actual entries first to put it closer to the previous two lines,
+	// which have the most important expected entries.
+	failureLines = append(failureLines, "Actual:")
+	for i, line := range logLines {
+		failureLines = append(failureLines, fmt.Sprintf("  [%d]: %q", i, line))
+	}
+	// Then give all the expected entries for good form.
+	failureLines = append(failureLines, "Expected:")
+	for i, line := range expInLog {
+		failureLines = append(failureLines, fmt.Sprintf("  [%d]: %q", i, line))
+	}
+	return s.Failf(strings.Join(failureLines, "\n"), msg, args...)
+}
+
+// assertLogDoesNotContain asserts that the log output does not contain any of the expNotInLog as substrings.
+// Designed for AssertLogContents, please use that.
+func (s *UpgradeTestSuite) assertLogDoesNotContain(logOutput string, expNotInLog []string, msg string, args ...interface{}) bool {
+	noneThere := true
 	for _, unexp := range expNotInLog {
-		// Same here with the msgAndArgs thing.
-		s.Assert().NotContains(logOutput, unexp, "upgrades[%q].Handler(...) log output.\nNot Expecting: %q", key, unexp)
+		// I'm including the unexpected in the msgAndArgs here even though it's also included in the
+		// failure message because the failure message puts everything on one line. So the expected
+		// string is almost at the end of a very long line. Having it in the "message" portion of
+		// the failure makes it easier to identify the problematic entry.
+		if !s.Assert().NotContainsf(logOutput, unexp, msg+"\nNot Expecting: %q", append(args, unexp)...) {
+			noneThere = false
+		}
 	}
+	return noneThere
+}
 
-	return
+// CreateAndFundAccount creates a new account in the app and funds it with the provided coin.
+func (s *UpgradeTestSuite) CreateAndFundAccount(coin sdk.Coin) sdk.AccAddress {
+	key2 := secp256k1.GenPrivKey()
+	pub2 := key2.PubKey()
+	addr2 := sdk.AccAddress(pub2.Address())
+	s.LogIfError(testutil.FundAccount(s.app.BankKeeper, s.ctx, addr2, sdk.Coins{coin}), "FundAccount(..., %q)", coin.String())
+	return addr2
+}
+
+// CreateValidator creates a new validator in the app.
+func (s *UpgradeTestSuite) CreateValidator(unbondedTime time.Time, status stakingtypes.BondStatus) stakingtypes.Validator {
+	key := secp256k1.GenPrivKey()
+	pub := key.PubKey()
+	addr := sdk.AccAddress(pub.Address())
+	valAddr := sdk.ValAddress(addr)
+	validator, err := stakingtypes.NewValidator(valAddr, pub, stakingtypes.NewDescription(valAddr.String(), "", "", "", ""))
+	s.Require().NoError(err, "could not init new validator")
+	validator.UnbondingTime = unbondedTime
+	validator.Status = status
+	s.app.StakingKeeper.SetValidator(s.ctx, validator)
+	err = s.app.StakingKeeper.SetValidatorByConsAddr(s.ctx, validator)
+	s.Require().NoError(err, "could not SetValidatorByConsAddr ")
+	err = s.app.StakingKeeper.AfterValidatorCreated(s.ctx, validator.GetOperator())
+	s.Require().NoError(err, "could not AfterValidatorCreated")
+	return validator
+}
+
+// DelegateToValidator delegates to a validator in the app.
+func (s *UpgradeTestSuite) DelegateToValidator(valAddress sdk.ValAddress, delegatorAddress sdk.AccAddress, coin sdk.Coin) {
+	validator, found := s.app.StakingKeeper.GetValidator(s.ctx, valAddress)
+	s.Require().True(found, "GetValidator(%q)", valAddress.String())
+	_, err := s.app.StakingKeeper.Delegate(s.ctx, delegatorAddress, coin.Amount, types.Unbonded, validator, true)
+	s.Require().NoError(err, "Delegate(%q, %q, unbonded, %q, true) error", delegatorAddress.String(), coin.String(), valAddress.String())
 }
 
 func (s *UpgradeTestSuite) TestKeysInHandlersMap() {
@@ -232,6 +387,7 @@ func (s *UpgradeTestSuite) TestRustRC1() {
 
 	expInLog := []string{
 		"INF Starting module migrations. This may take a significant amount of time to complete. Do not restart node.",
+		"INF removing all delegations from validators that have been inactive (unbonded) for 21 days",
 		`INF Setting "accountdata" name record.`,
 		`INF Creating message fee for "/cosmos.gov.v1.MsgSubmitProposal" if it doesn't already exist.`,
 		`INF Removing message fee for "/provenance.metadata.v1.MsgP8eMemorializeContractRequest" if one exists.`,
@@ -248,6 +404,7 @@ func (s *UpgradeTestSuite) TestRust() {
 
 	expInLog := []string{
 		"INF Starting module migrations. This may take a significant amount of time to complete. Do not restart node.",
+		"INF removing all delegations from validators that have been inactive (unbonded) for 21 days",
 		`INF Setting "accountdata" name record.`,
 		`INF Removing message fee for "/provenance.metadata.v1.MsgP8eMemorializeContractRequest" if one exists.`,
 		"INF Fixing name module store index entries.",
@@ -259,13 +416,213 @@ func (s *UpgradeTestSuite) TestRust() {
 	s.AssertUpgradeHandlerLogs("rust", expInLog, expNotInLog)
 }
 
+func (s *UpgradeTestSuite) TestRemoveInactiveValidatorDelegations() {
+	addr1 := s.CreateAndFundAccount(sdk.NewInt64Coin("stake", 1000000))
+	addr2 := s.CreateAndFundAccount(sdk.NewInt64Coin("stake", 1000000))
+
+	runner := func() {
+		removeInactiveValidatorDelegations(s.ctx, s.app)
+	}
+	runnerName := "removeInactiveValidatorDelegations"
+
+	delegationCoin := sdk.NewInt64Coin("stake", 10000)
+	delegationCoinAmt := sdk.NewDec(delegationCoin.Amount.Int64())
+
+	s.Run("just one bonded validator", func() {
+		validators := s.app.StakingKeeper.GetAllValidators(s.ctx)
+		s.Require().Len(validators, 1, "GetAllValidators after setup")
+
+		expectedLogLines := []string{
+			"INF removing all delegations from validators that have been inactive (unbonded) for 21 days",
+			"INF a total of 0 inactive (unbonded) validators have had all their delegators removed",
+		}
+		s.ExecuteAndAssertLogs(runner, expectedLogLines, nil, true, runnerName)
+
+		newValidators := s.app.StakingKeeper.GetAllValidators(s.ctx)
+		s.Require().Len(newValidators, 1, "GetAllValidators after %s", runnerName)
+	})
+
+	s.Run("one unbonded validator with a delegation", func() {
+		// single unbonded validator with 1 delegations, should be removed
+		unbondedVal1 := s.CreateValidator(s.startTime.Add(-30*24*time.Hour), stakingtypes.Unbonded)
+		addr1Balance := s.app.BankKeeper.GetBalance(s.ctx, addr1, "stake")
+		s.DelegateToValidator(unbondedVal1.GetOperator(), addr1, delegationCoin)
+		s.Require().Equal(addr1Balance.Sub(delegationCoin), s.app.BankKeeper.GetBalance(s.ctx, addr1, "stake"), "addr1 should have less funds")
+		validators := s.app.StakingKeeper.GetAllValidators(s.ctx)
+		s.Require().Len(validators, 2, "Setup: GetAllValidators should have: 1 bonded, 1 unbonded")
+
+		expectedLogLines := []string{
+			"INF removing all delegations from validators that have been inactive (unbonded) for 21 days",
+			fmt.Sprintf("INF validator %v has been inactive (unbonded) for %d days and will be removed", unbondedVal1.OperatorAddress, 30),
+			fmt.Sprintf("INF undelegate delegator %v from validator %v of all shares (%v)", addr1.String(), unbondedVal1.OperatorAddress, delegationCoinAmt),
+			"INF a total of 1 inactive (unbonded) validators have had all their delegators removed",
+		}
+		s.ExecuteAndAssertLogs(runner, expectedLogLines, nil, true, runnerName)
+
+		validators = s.app.StakingKeeper.GetAllValidators(s.ctx)
+		s.Assert().Len(validators, 1, "GetAllValidators after %s", runnerName)
+		ubd, found := s.app.StakingKeeper.GetUnbondingDelegation(s.ctx, addr1, unbondedVal1.GetOperator())
+		s.Assert().True(found, "GetUnbondingDelegation found")
+		if s.Assert().Len(ubd.Entries, 1, "UnbondingDelegation entries") {
+			s.Assert().Equal(delegationCoin.Amount, ubd.Entries[0].Balance, "UnbondingDelegation balance")
+		}
+	})
+
+	s.Run("one unbonded validator with 2 delegations", func() {
+		// single unbonded validator with 2 delegations
+		unbondedVal1 := s.CreateValidator(s.startTime.Add(-30*24*time.Hour), stakingtypes.Unbonded)
+		addr1Balance := s.app.BankKeeper.GetBalance(s.ctx, addr1, "stake")
+		addr2Balance := s.app.BankKeeper.GetBalance(s.ctx, addr2, "stake")
+		s.DelegateToValidator(unbondedVal1.GetOperator(), addr1, delegationCoin)
+		s.DelegateToValidator(unbondedVal1.GetOperator(), addr2, delegationCoin)
+		s.Require().Equal(addr1Balance.Sub(delegationCoin), s.app.BankKeeper.GetBalance(s.ctx, addr1, "stake"), "addr1 should have less funds after delegation")
+		s.Require().Equal(addr2Balance.Sub(delegationCoin), s.app.BankKeeper.GetBalance(s.ctx, addr2, "stake"), "addr2 should have less funds after delegation")
+		var found bool
+		unbondedVal1, found = s.app.StakingKeeper.GetValidator(s.ctx, unbondedVal1.GetOperator())
+		s.Require().True(found, "Setup: GetValidator(unbondedVal1) found")
+		s.Require().Equal(delegationCoinAmt.Add(delegationCoinAmt), unbondedVal1.DelegatorShares, "Setup: unbondedVal1.DelegatorShares")
+		validators := s.app.StakingKeeper.GetAllValidators(s.ctx)
+		s.Require().Len(validators, 2, "Setup: GetAllValidators should have: 1 bonded, 1 unbonded")
+
+		expectedLogLines := []string{
+			"INF removing all delegations from validators that have been inactive (unbonded) for 21 days",
+			fmt.Sprintf("INF validator %v has been inactive (unbonded) for %d days and will be removed", unbondedVal1.OperatorAddress, 30),
+			fmt.Sprintf("INF undelegate delegator %v from validator %v of all shares (%v)", addr1.String(), unbondedVal1.OperatorAddress, delegationCoinAmt),
+			fmt.Sprintf("INF undelegate delegator %v from validator %v of all shares (%v)", addr2.String(), unbondedVal1.OperatorAddress, delegationCoinAmt),
+			"INF a total of 1 inactive (unbonded) validators have had all their delegators removed",
+		}
+		s.ExecuteAndAssertLogs(runner, expectedLogLines, nil, false, runnerName)
+
+		validators = s.app.StakingKeeper.GetAllValidators(s.ctx)
+		s.Assert().Len(validators, 1, "GetAllValidators after %s", runnerName)
+		ubd, found := s.app.StakingKeeper.GetUnbondingDelegation(s.ctx, addr1, unbondedVal1.GetOperator())
+		s.Assert().True(found, "GetUnbondingDelegation(addr1) found")
+		if s.Assert().Len(ubd.Entries, 1, "GetUnbondingDelegation(addr1) entries") {
+			s.Assert().Equal(delegationCoin.Amount, ubd.Entries[0].Balance, "GetUnbondingDelegation(addr1) balance")
+		}
+		ubd, found = s.app.StakingKeeper.GetUnbondingDelegation(s.ctx, addr2, unbondedVal1.GetOperator())
+		s.Assert().True(found, "GetUnbondingDelegation(addr2) found")
+		if s.Assert().Len(ubd.Entries, 1, "GetUnbondingDelegation(addr2) entries") {
+			s.Assert().Equal(delegationCoin.Amount, ubd.Entries[0].Balance, "GetUnbondingDelegation(addr2) balance")
+		}
+	})
+
+	s.Run("two unbonded validators to be removed", func() {
+		// 2 unbonded validators with delegations past inactive time, both should be removed
+		unbondedVal1 := s.CreateValidator(s.startTime.Add(-30*24*time.Hour), stakingtypes.Unbonded)
+		s.DelegateToValidator(unbondedVal1.GetOperator(), addr1, delegationCoin)
+		s.DelegateToValidator(unbondedVal1.GetOperator(), addr2, delegationCoin)
+		var found bool
+		unbondedVal1, found = s.app.StakingKeeper.GetValidator(s.ctx, unbondedVal1.GetOperator())
+		s.Require().True(found, "Setup: GetValidator(unbondedVal1) found")
+		s.Require().Equal(delegationCoinAmt.Add(delegationCoinAmt), unbondedVal1.DelegatorShares, "Setup: shares delegated to unbondedVal1")
+
+		unbondedVal2 := s.CreateValidator(s.startTime.Add(-29*24*time.Hour), stakingtypes.Unbonded)
+		s.DelegateToValidator(unbondedVal2.GetOperator(), addr1, delegationCoin)
+		s.DelegateToValidator(unbondedVal2.GetOperator(), addr2, delegationCoin)
+		unbondedVal2, found = s.app.StakingKeeper.GetValidator(s.ctx, unbondedVal2.GetOperator())
+		s.Require().True(found, "Setup: GetValidator(unbondedVal2) found")
+		s.Require().Equal(delegationCoinAmt.Add(delegationCoinAmt), unbondedVal2.DelegatorShares, "Setup: shares delegated to unbondedVal2")
+		validators := s.app.StakingKeeper.GetAllValidators(s.ctx)
+		s.Require().Len(validators, 3, "Setup: GetAllValidators should have: 1 bonded, 2 unbonded")
+
+		expectedLogLines := []string{
+			"INF removing all delegations from validators that have been inactive (unbonded) for 21 days",
+			fmt.Sprintf("INF validator %v has been inactive (unbonded) for %d days and will be removed", unbondedVal1.OperatorAddress, 30),
+			fmt.Sprintf("INF undelegate delegator %v from validator %v of all shares (%v)", addr1.String(), unbondedVal1.OperatorAddress, delegationCoinAmt),
+			fmt.Sprintf("INF undelegate delegator %v from validator %v of all shares (%v)", addr2.String(), unbondedVal1.OperatorAddress, delegationCoinAmt),
+			fmt.Sprintf("INF validator %v has been inactive (unbonded) for %d days and will be removed", unbondedVal2.OperatorAddress, 29),
+			fmt.Sprintf("INF undelegate delegator %v from validator %v of all shares (%v)", addr1.String(), unbondedVal2.OperatorAddress, delegationCoinAmt),
+			fmt.Sprintf("INF undelegate delegator %v from validator %v of all shares (%v)", addr2.String(), unbondedVal2.OperatorAddress, delegationCoinAmt),
+			"INF a total of 2 inactive (unbonded) validators have had all their delegators removed",
+		}
+		s.ExecuteAndAssertLogs(runner, expectedLogLines, nil, false, runnerName)
+
+		validators = s.app.StakingKeeper.GetAllValidators(s.ctx)
+		s.Assert().Len(validators, 1, "GetAllValidators after %s", runnerName)
+		ubd, found := s.app.StakingKeeper.GetUnbondingDelegation(s.ctx, addr1, unbondedVal1.GetOperator())
+		s.Assert().True(found, "GetUnbondingDelegation(addr1) found")
+		if s.Assert().Len(ubd.Entries, 1, "GetUnbondingDelegation(addr1) entries") {
+			s.Assert().Equal(delegationCoin.Amount, ubd.Entries[0].Balance, "GetUnbondingDelegation(addr1) balance")
+		}
+		ubd, found = s.app.StakingKeeper.GetUnbondingDelegation(s.ctx, addr2, unbondedVal1.GetOperator())
+		s.Assert().True(found, "GetUnbondingDelegation(addr2) found")
+		if s.Assert().Len(ubd.Entries, 1, "GetUnbondingDelegation(addr2) entries") {
+			s.Assert().Equal(delegationCoin.Amount, ubd.Entries[0].Balance, "GetUnbondingDelegation(addr2) balance")
+		}
+	})
+
+	s.Run("two unbonded validators one too recently", func() {
+		// 2 unbonded validators, 1 under the inactive day count, should only remove one
+		unbondedVal1 := s.CreateValidator(s.startTime.Add(-30*24*time.Hour), stakingtypes.Unbonded)
+		s.DelegateToValidator(unbondedVal1.GetOperator(), addr1, delegationCoin)
+		s.DelegateToValidator(unbondedVal1.GetOperator(), addr2, delegationCoin)
+		var found bool
+		unbondedVal1, found = s.app.StakingKeeper.GetValidator(s.ctx, unbondedVal1.GetOperator())
+		s.Require().True(found, "Setup: GetValidator(unbondedVal1) found")
+		s.Require().Equal(delegationCoinAmt.Add(delegationCoinAmt), unbondedVal1.DelegatorShares, "Setup: shares delegated to unbondedVal1")
+
+		unbondedVal2 := s.CreateValidator(s.startTime.Add(-20*24*time.Hour), stakingtypes.Unbonded)
+		s.DelegateToValidator(unbondedVal2.GetOperator(), addr1, delegationCoin)
+		unbondedVal2, found = s.app.StakingKeeper.GetValidator(s.ctx, unbondedVal2.GetOperator())
+		s.Require().True(found, "Setup: GetValidator(unbondedVal2) found")
+		s.Require().Equal(delegationCoinAmt, unbondedVal2.DelegatorShares, "Setup: shares delegated to unbondedVal1")
+		validators := s.app.StakingKeeper.GetAllValidators(s.ctx)
+		s.Require().Len(validators, 3, "Setup: GetAllValidators should have: 1 bonded, 1 recently unbonded, 1 old unbonded")
+
+		expectedLogLines := []string{
+			"INF removing all delegations from validators that have been inactive (unbonded) for 21 days",
+			fmt.Sprintf("INF validator %v has been inactive (unbonded) for %d days and will be removed", unbondedVal1.OperatorAddress, 30),
+			fmt.Sprintf("INF undelegate delegator %v from validator %v of all shares (%v)", addr1.String(), unbondedVal1.OperatorAddress, delegationCoinAmt),
+			fmt.Sprintf("INF undelegate delegator %v from validator %v of all shares (%v)", addr2.String(), unbondedVal1.OperatorAddress, delegationCoinAmt),
+			"INF a total of 1 inactive (unbonded) validators have had all their delegators removed",
+		}
+		notExpectedLogLines := []string{
+			fmt.Sprintf("validator %v has been inactive (unbonded)", unbondedVal2.OperatorAddress),
+			fmt.Sprintf("undelegate delegator %v from validator %v", addr1.String(), unbondedVal2.OperatorAddress),
+			fmt.Sprintf("undelegate delegator %v from validator %v", addr2.String(), unbondedVal2.OperatorAddress),
+		}
+		s.ExecuteAndAssertLogs(runner, expectedLogLines, notExpectedLogLines, false, runnerName)
+
+		validators = s.app.StakingKeeper.GetAllValidators(s.ctx)
+		s.Assert().Len(validators, 2, "GetAllValidators after %s", runnerName)
+		ubd, found := s.app.StakingKeeper.GetUnbondingDelegation(s.ctx, addr1, unbondedVal1.GetOperator())
+		s.Assert().True(found, "GetUnbondingDelegation(addr1) found")
+		if s.Assert().Len(ubd.Entries, 1, "GetUnbondingDelegation(addr1) entries") {
+			s.Assert().Equal(delegationCoin.Amount, ubd.Entries[0].Balance, "GetUnbondingDelegation(addr1) balance")
+		}
+		ubd, found = s.app.StakingKeeper.GetUnbondingDelegation(s.ctx, addr2, unbondedVal1.GetOperator())
+		s.Assert().True(found, "GetUnbondingDelegation(addr2) found")
+		if s.Assert().Len(ubd.Entries, 1, "GetUnbondingDelegation(addr2) entries") {
+			s.Assert().Equal(delegationCoin.Amount, ubd.Entries[0].Balance, "GetUnbondingDelegation(addr2) balance")
+		}
+	})
+
+	s.Run("unbonded without delegators", func() {
+		// create a unbonded validator with out delegators, should not remove
+		unbondedVal1 := s.CreateValidator(s.startTime.Add(-30*24*time.Hour), stakingtypes.Unbonded)
+		validators := s.app.StakingKeeper.GetAllValidators(s.ctx)
+		s.Require().Len(validators, 3, "Setup: GetAllValidators should have: 1 bonded, 1 recently unbonded, 1 empty unbonded")
+
+		expectedLogLines := []string{
+			"INF removing all delegations from validators that have been inactive (unbonded) for 21 days",
+			fmt.Sprintf("INF validator %v has been inactive (unbonded) for %d days and will be removed", unbondedVal1.OperatorAddress, 30),
+			"INF a total of 1 inactive (unbonded) validators have had all their delegators removed",
+		}
+		s.ExecuteAndAssertLogs(runner, expectedLogLines, nil, true, runnerName)
+
+		validators = s.app.StakingKeeper.GetAllValidators(s.ctx)
+		s.Assert().Len(validators, 3, "GetAllValidators after %s", runnerName)
+	})
+}
+
 func (s *UpgradeTestSuite) TestAddGovV1SubmitFee() {
 	v1TypeURL := "/cosmos.gov.v1.MsgSubmitProposal"
 	v1B1TypeURL := "/cosmos.gov.v1beta1.MsgSubmitProposal"
 
-	startingMsg := `Creating message fee for "` + v1TypeURL + `" if it doesn't already exist.`
+	startingMsg := `INF Creating message fee for "` + v1TypeURL + `" if it doesn't already exist.`
 	successMsg := func(amt string) string {
-		return `Successfully set fee for "` + v1TypeURL + `" with amount "` + amt + `".`
+		return `INF Successfully set fee for "` + v1TypeURL + `" with amount "` + amt + `".`
 	}
 
 	coin := func(denom string, amt int64) *sdk.Coin {
@@ -286,7 +643,7 @@ func (s *UpgradeTestSuite) TestAddGovV1SubmitFee() {
 			v1B1Amt: coin("betacoin", 99),
 			expInLog: []string{
 				startingMsg,
-				`Message fee for "` + v1TypeURL + `" already exists with amount "88foocoin". Nothing to do.`,
+				`INF Message fee for "` + v1TypeURL + `" already exists with amount "88foocoin". Nothing to do.`,
 			},
 			expAmt: *coin("foocoin", 88),
 		},
@@ -295,7 +652,7 @@ func (s *UpgradeTestSuite) TestAddGovV1SubmitFee() {
 			v1B1Amt: coin("betacoin", 99),
 			expInLog: []string{
 				startingMsg,
-				`Copying "` + v1B1TypeURL + `" fee to "` + v1TypeURL + `".`,
+				`INF Copying "` + v1B1TypeURL + `" fee to "` + v1TypeURL + `".`,
 				successMsg("99betacoin"),
 			},
 			expAmt: *coin("betacoin", 99),
@@ -304,7 +661,7 @@ func (s *UpgradeTestSuite) TestAddGovV1SubmitFee() {
 			name: "brand new",
 			expInLog: []string{
 				startingMsg,
-				`Creating "` + v1TypeURL + `" fee.`,
+				`INF Creating "` + v1TypeURL + `" fee.`,
 				successMsg("100000000000nhash"),
 			},
 			expAmt: *coin("nhash", 100_000_000_000),
@@ -341,14 +698,14 @@ func (s *UpgradeTestSuite) TestAddGovV1SubmitFee() {
 			testFunc := func() {
 				addGovV1SubmitFee(s.ctx, s.app)
 			}
-			s.Require().NotPanics(testFunc, "addGovV1SubmitFee")
-			logOutput := s.logBuffer.String()
-			s.T().Logf("addGovV1SubmitFee log output:\n%s", logOutput)
+			didNotPanic := s.Assert().NotPanics(testFunc, "addGovV1SubmitFee")
+			logOutput := s.GetLogOutput("addGovV1SubmitFee")
+			if !didNotPanic {
+				return
+			}
 
 			// Make sure the log has the expected lines.
-			for _, exp := range tc.expInLog {
-				s.Assert().Contains(logOutput, exp, "addGovV1SubmitFee log output")
-			}
+			s.AssertLogContents(logOutput, tc.expInLog, nil, true, "addGovV1SubmitFee")
 
 			// Get the fee and make sure it's now as expected.
 			fee, err := s.app.MsgFeesKeeper.GetMsgFee(s.ctx, v1TypeURL)
@@ -362,7 +719,7 @@ func (s *UpgradeTestSuite) TestAddGovV1SubmitFee() {
 
 func (s *UpgradeTestSuite) TestRemoveP8eMemorializeContractFee() {
 	typeURL := "/provenance.metadata.v1.MsgP8eMemorializeContractRequest"
-	startingMsg := `Removing message fee for "` + typeURL + `" if one exists.`
+	startingMsg := `INF Removing message fee for "` + typeURL + `" if one exists.`
 
 	coin := func(denom string, amt int64) *sdk.Coin {
 		rv := sdk.NewInt64Coin(denom, amt)
@@ -378,7 +735,7 @@ func (s *UpgradeTestSuite) TestRemoveP8eMemorializeContractFee() {
 			name: "does not exist",
 			expInLog: []string{
 				startingMsg,
-				`Message fee for "` + typeURL + `" already does not exist. Nothing to do.`,
+				`INF Message fee for "` + typeURL + `" already does not exist. Nothing to do.`,
 			},
 		},
 		{
@@ -386,7 +743,7 @@ func (s *UpgradeTestSuite) TestRemoveP8eMemorializeContractFee() {
 			amt:  coin("p8ecoin", 808),
 			expInLog: []string{
 				startingMsg,
-				`Successfully removed message fee for "` + typeURL + `" with amount "808p8ecoin".`,
+				`INF Successfully removed message fee for "` + typeURL + `" with amount "808p8ecoin".`,
 			},
 		},
 	}
@@ -404,17 +761,19 @@ func (s *UpgradeTestSuite) TestRemoveP8eMemorializeContractFee() {
 				}
 			}
 
+			// Reset the log buffer to clear out unrelated entries.
 			s.logBuffer.Reset()
+			// Call removeP8eMemorializeContractFee and relog its output (to help if things fail).
 			testFunc := func() {
 				removeP8eMemorializeContractFee(s.ctx, s.app)
 			}
-			s.Require().NotPanics(testFunc, "removeP8eMemorializeContractFee")
-			logOutput := s.logBuffer.String()
-			s.T().Logf("removeP8eMemorializeContractFee log output:\n%s", logOutput)
-
-			for _, exp := range tc.expInLog {
-				s.Assert().Contains(logOutput, exp, "removeP8eMemorializeContractFee log output")
+			didNotPanic := s.Assert().NotPanics(testFunc, "removeP8eMemorializeContractFee")
+			logOutput := s.GetLogOutput("removeP8eMemorializeContractFee")
+			if !didNotPanic {
+				return
 			}
+
+			s.AssertLogContents(logOutput, tc.expInLog, nil, true, "removeP8eMemorializeContractFee")
 
 			// Make sure there isn't a fee anymore.
 			fee, err := s.app.MsgFeesKeeper.GetMsgFee(s.ctx, typeURL)
@@ -425,25 +784,29 @@ func (s *UpgradeTestSuite) TestRemoveP8eMemorializeContractFee() {
 }
 
 func (s *UpgradeTestSuite) TestSetAccountDataNameRecord() {
-	// Since this is also done during InitGenesis, it should already be set up as needed.
+	// Most of the testing should (hopefully) be done in the name module. So this is
+	// just a superficial test that makes sure it's doing something.
+	// Since the name is also created during InitGenesis, it should already be set up as needed.
 	// So in this unit test, the logs will indicate that.
-	expLogLines := []string{
+	expInLog := []string{
 		`INF Setting "accountdata" name record.`,
 		`INF The "accountdata" name record already exists as needed. Nothing to do.`,
-		"", // The log lines all end with a \n. So when I split it on \n there will be an empty string at the end.
 	}
 	// During an actual upgrade, that last line would instead be this:
 	// `INF Successfully set "accountdata" name record.`
 
+	// Reset the log buffer to clear out unrelated entries.
 	s.logBuffer.Reset()
+	// Call setAccountDataNameRecord and relog its output (to help if things fail).
 	var err error
 	testFunc := func() {
 		err = setAccountDataNameRecord(s.ctx, s.app.AccountKeeper, &s.app.NameKeeper)
 	}
-	s.Require().NotPanics(testFunc, "setAccountDataNameRecord")
-	logged := s.logBuffer.String()
-	s.T().Logf("Logs generated during setAccountDataNameRecord:\n%s", logged)
+	didNotPanic := s.Assert().NotPanics(testFunc, "setAccountDataNameRecord")
+	logOutput := s.GetLogOutput("setAccountDataNameRecord")
+	if !didNotPanic {
+		return
+	}
 	s.Require().NoError(err, "setAccountDataNameRecord")
-	loggedLines := strings.Split(logged, "\n")
-	s.Assert().Equal(expLogLines, loggedLines, "lines logged during setAccountDataNameRecord")
+	s.AssertLogContents(logOutput, expInLog, nil, true, "setAccountDataNameRecord")
 }