+++ CHANGELOG.md
@@ -60,6 +60,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Bumped ibc-go to 6.1.1 [#1563](https://github.com/provenance-io/provenance/pull/1563).
 * Update `marker` module spec documentation with new proto references [#1580](https://github.com/provenance-io/provenance/pull/1580).
 * Bumped `wasmd` to v0.30.0-pio-5 and `wasmvm` to v1.2.4 [#1582](https://github.com/provenance-io/provenance/pull/1582).
+* Inactive validator delegation cleanup process [#1556](https://github.com/provenance-io/provenance/issues/1556).
 * Bump Cosmos-SDK to [v0.46.13-pio-1](https://github.com/provenance-io/cosmos-sdk/blob/v0.46.13-pio-1/RELEASE_NOTES.md) (from `v0.46.10-pio-4`) [PR 1585](https://github.com/provenance-io/provenance/pull/1585).
 
 ### Bug Fixes